--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "test-data/gardenlinux"]
-	path = test-data/gardenlinux
-	url = https://github.com/gardenlinux/gardenlinux
