--- example-data/extras
@@ -1,1 +1,0 @@
-extrastuff
