--- example-data/foo.bar
@@ -1,1 +1,0 @@
-foobar
