--- example-data/foo2.bar
@@ -1,1 +1,0 @@
-different foobar 2
