--- requirements.txt
@@ -1,45 +1,0 @@
-alabaster==0.7.16 ; python_version >= "3.12" and python_version < "4.0"
-apt-repo==0.5 ; python_version >= "3.12" and python_version < "4.0"
-attrs==24.2.0 ; python_version >= "3.12" and python_version < "4.0"
-babel==2.16.0 ; python_version >= "3.12" and python_version < "4.0"
-certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
-cffi==1.17.0 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
-charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
-click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
-colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
-cryptography==43.0.1 ; python_version >= "3.12" and python_version < "4.0"
-docutils==0.20.1 ; python_version >= "3.12" and python_version < "4.0"
-gitdb==4.0.11 ; python_version >= "3.12" and python_version < "4.0"
-gitpython==3.1.43 ; python_version >= "3.12" and python_version < "4.0"
-idna==3.8 ; python_version >= "3.12" and python_version < "4.0"
-imagesize==1.4.1 ; python_version >= "3.12" and python_version < "4.0"
-iniconfig==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
-jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
-jsonschema-specifications==2023.12.1 ; python_version >= "3.12" and python_version < "4.0"
-jsonschema==4.23.0 ; python_version >= "3.12" and python_version < "4.0"
-markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
-networkx==3.3 ; python_version >= "3.12" and python_version < "4.0"
-oras @ git+https://github.com/oras-project/oras-py.git@caf8db5b279382335fbb1f6d7402ed9b73618d37 ; python_version >= "3.12" and python_version < "4.0"
-packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
-pluggy==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
-pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
-pygments==2.18.0 ; python_version >= "3.12" and python_version < "4.0"
-pytest==8.3.2 ; python_version >= "3.12" and python_version < "4.0"
-python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
-python-gardenlinux-lib @ git+https://github.com/gardenlinux/python-gardenlinux-lib@fe882e7a136ccd711e9ea274204c5d7bba587554 ; python_version >= "3.12" and python_version < "4.0"
-pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
-referencing==0.35.1 ; python_version >= "3.12" and python_version < "4.0"
-requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
-rpds-py==0.20.0 ; python_version >= "3.12" and python_version < "4.0"
-smmap==5.0.1 ; python_version >= "3.12" and python_version < "4.0"
-snowballstemmer==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
-sphinx-rtd-theme==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
-sphinx==7.4.7 ; python_version >= "3.12" and python_version < "4.0"
-sphinxcontrib-applehelp==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
-sphinxcontrib-devhelp==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
-sphinxcontrib-htmlhelp==2.1.0 ; python_version >= "3.12" and python_version < "4.0"
-sphinxcontrib-jquery==4.1 ; python_version >= "3.12" and python_version < "4.0"
-sphinxcontrib-jsmath==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
-sphinxcontrib-qthelp==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
-sphinxcontrib-serializinghtml==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
-urllib3==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
