+++ src/glcli.py
@@ -1,8 +1,6 @@
 #!/bin/env python3
 
 import click
-from ocm import ocm
-from apt import apt
 from oci import oci
 
 
@@ -15,9 +13,6 @@ def glcli():
 
 
 glcli.add_command(oci)
-glcli.add_command(ocm)
-glcli.add_command(apt)
-
 
 if __name__ == "__main__":
     glcli()
